+++ src/views/ImprintPage.vue
@@ -16,9 +16,9 @@
             <div>
               <h2 class="text-2xl font-semibold mb-4">Angaben gemäß § 5 TMG</h2>
               <p class="text-gray-700">
-                MedMatch GmbH<br>
-                Musterstraße 123<br>
-                10115 Berlin<br>
+                Medmatch<br>
+                Hülshoffstr.22<br>
+                46342 Velen<br>
                 Deutschland
               </p>
             </div>